pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "hh-vacancies-project"
version = "1.0.0"
description = "Project for working with HH.ru vacancies database"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "hh_vacancies", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.8"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = ">=5.0.0,<7.0.0"
isort = "^5.12.0"

[tool.black]
line-length = 119
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 119

[tool.flake8]
max-line-length = 119
extend-ignore = "E203,W503"
exclude = ".git,__pycache__,build,dist,.venv,venv"